Revert "Create init payload capability for Rich and Plain Text (#1217)"
This reverts commit 9b237aa4a38074b26facc0483a4b4a2d62d34349.
