+++ packages/lexical-playground/src/Editor.js
@@ -7,8 +7,6 @@
  * @flow strict
  */
 
-import type {LexicalEditor, RootNode} from 'lexical';
-
 import * as React from 'react';
 
 import PlainTextPlugin from '@lexical/react/LexicalPlainTextPlugin';
@@ -41,44 +39,11 @@ import {useSharedHistoryContext} from './context/SharedHistoryContext';
 import ContentEditable from './ui/ContentEditable';
 import AutoLinkPlugin from './plugins/AutoLinkPlugin';
 import PollPlugin from './plugins/PollPlugin';
-
-import {$getRoot, $getSelection} from 'lexical';
-import {$createParagraphNode} from 'lexical/ParagraphNode';
 import {useSettings} from './context/SettingsContext';
 import AutoFocusPlugin from './plugins/AutoFocusPlugin';
 
 const skipCollaborationInit =
   window.parent != null && window.parent.frames.right === window;
-
-function shouldSelectParagraph(editor: LexicalEditor): boolean {
-  const activeElement = document.activeElement;
-  return (
-    $getSelection() !== null ||
-    (activeElement !== null && activeElement === editor.getRootElement())
-  );
-}
-
-function initParagraph(root: RootNode, editor: LexicalEditor): void {
-  const paragraph = $createParagraphNode();
-  root.append(paragraph);
-  if (shouldSelectParagraph(editor)) {
-    paragraph.select();
-  }
-}
-
-function textInitFn(editor: LexicalEditor): void {
-  const root = $getRoot();
-  const firstChild = root.getFirstChild();
-  if (firstChild === null) {
-    initParagraph(root, editor);
-  }
-}
-
-function clearEditor(editor: LexicalEditor): void {
-  const root = $getRoot();
-  root.clear();
-  initParagraph(root, editor);
-}
 
 export default function Editor(): React$Node {
   const {historyState} = useSharedHistoryContext();
@@ -130,8 +95,7 @@ export default function Editor(): React$Node {
             <RichTextPlugin
               contentEditable={<ContentEditable />}
               placeholder={placeholder}
-              initialPayloadFn={textInitFn}
-              clearEditorFn={clearEditor}
+              skipInit={isCollab}
             />
             <AutoFormatterPlugin />
             <CodeHighlightPlugin />
@@ -147,8 +111,6 @@ export default function Editor(): React$Node {
             <PlainTextPlugin
               contentEditable={<ContentEditable />}
               placeholder={placeholder}
-              initialPayloadFn={textInitFn}
-              clearEditorFn={clearEditor}
             />
             <HistoryPlugin externalHistoryState={historyState} />
           </>
