+++ packages/lexical-react/src/LexicalPlainTextPlugin.js
@@ -7,8 +7,6 @@
  * @flow strict
  */
 
-import type {LexicalEditor} from 'lexical';
-
 import * as React from 'react';
 
 import {useLexicalComposerContext} from '@lexical/react/LexicalComposerContext';
@@ -24,19 +22,17 @@ function onErrorDefault(e: Error): void {
 export default function PlainTextPlugin({
   contentEditable,
   placeholder,
-  initialPayloadFn,
-  clearEditorFn,
+  skipInit,
   onError,
 }: {
   contentEditable: React$Node,
   placeholder: React$Node,
-  initialPayloadFn?: (LexicalEditor) => void,
-  clearEditorFn?: (LexicalEditor) => void,
+  skipInit?: boolean,
   onError?: (error: Error, log: Array<string>) => void,
 }): React$Node {
   const [editor] = useLexicalComposerContext();
   const showPlaceholder = useCanShowPlaceholder(editor);
-  usePlainTextSetup(editor, initialPayloadFn, clearEditorFn);
+  usePlainTextSetup(editor, !skipInit);
   const decorators = useDecorators(editor);
 
   useLayoutEffect(() => {
