+++ packages/lexical-react/src/shared/useYjsCollaboration.js
@@ -7,11 +7,7 @@
  * @flow strict
  */
 
-import type {
-  LexicalEditor,
-  RootNode,
-  CommandListenerEditorPriority,
-} from 'lexical';
+import type {LexicalEditor, CommandListenerEditorPriority} from 'lexical';
 import type {Provider, Binding} from '@lexical/yjs';
 import type {Doc} from 'yjs';
 
@@ -20,8 +16,6 @@ import * as React from 'react';
 import {createPortal} from 'react-dom';
 
 import {useCallback, useEffect, useMemo} from 'react';
-import {$getSelection, $getRoot} from 'lexical';
-import {$createParagraphNode} from 'lexical/ParagraphNode';
 import {
   createBinding,
   createUndoManager,
@@ -34,30 +28,6 @@ import {
 import {initEditor} from './useRichTextSetup';
 
 const EditorPriority: CommandListenerEditorPriority = 0;
-
-function shouldSelectParagraph(editor: LexicalEditor): boolean {
-  const activeElement = document.activeElement;
-  return (
-    $getSelection() !== null ||
-    (activeElement !== null && activeElement === editor.getRootElement())
-  );
-}
-
-function initParagraph(root: RootNode, editor: LexicalEditor): void {
-  const paragraph = $createParagraphNode();
-  root.append(paragraph);
-  if (shouldSelectParagraph(editor)) {
-    paragraph.select();
-  }
-}
-
-function richTextInitFn(editor: LexicalEditor): void {
-  const root = $getRoot();
-  const firstChild = root.getFirstChild();
-  if (firstChild === null) {
-    initParagraph(root, editor);
-  }
-}
 
 export function useYjsCollaboration(
   editor: LexicalEditor,
@@ -100,7 +70,7 @@ export function useYjsCollaboration(
         root.isEmpty() &&
         root._xmlText._length === 0
       ) {
-        initEditor(editor, richTextInitFn);
+        initEditor(editor);
       }
     };
